requirements.txt
Django=3.2.3
pytest=6.2.4
factory-boy=3.2.0
django-webtest=1.9.7